Fixed run, renamed few things, fixed bool from config
